Make blocks async, refactor
