+++ src/blocks/addMemeToTags.ts
@@ -1,7 +1,7 @@
 import { PullRequest } from '../types.js';
 const fs = require('fs');
 import path from 'path';
-export function addMemeToTags(pullRequest: PullRequest, config: { dbpath: string }) {
+export async function addMemeToTags(pullRequest: PullRequest, config: { dbpath: string }) {
   pullRequest.meme.tags.forEach((tagId) => {
     const jsonPath = path.join(config.dbpath, 'tags', `${tagId}.json`);
     let json;
