+++ tests/pr.test.ts
@@ -1,11 +1,11 @@
-import { run } from '../ci/pr-pipeline';
+import { run } from '../src/pr-pipeline';
 const path = require('path');
 
 const config = {
   dbpath: path.join(__dirname, '../db'),
 };
 
-it('can create new tags', () => {
-  const pr = run(path.join(__dirname, './fixtures/pr-with-new-tags.json'), config);
+it('can create new tags', async () => {
+  const pr = await run(path.join(__dirname, './fixtures/pr-with-new-tags.json'), config);
   expect(pr.meme.title).toBeTruthy;
 });
